Se puede modificar los datos del inicializados automáticamente

+++ Backend/src/main/java/com/CTA/UNLP/demo/repository/VehiculoRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface VehiculoRepository extends JpaRepository<Vehiculo,Long> {
-    public Optional<Vehiculo> findById(Long id);
+    public Optional<Vehiculo> findById(Integer id);
 }
